Make 1 the minimum weregeld payement
We believed there was a bug where paying weregeld when you had was free.
The actual bug was the person paying the weregeld still lost the minimum
of 1, but the person recveing there weregeld did not receive it. Now 1
geld is properly transferred even when the payer has no prestige.