Clean up TODO list
Changed the "todo" template at /todo. I removed the item to version
control the game since it has already been version controlled. I added a
header to the page. I removed the bug on the list which said that buying
lead viking didn't use up a slot. If the bug exists, it is clearly
intermittent. I think more likely someone was mistaken (ie, the track
cleared immediately because it was their turn next) or it was caused by
the update being non-atomic. If it was the latter, that is covered by
another bullet on the TODO list.