Make resource pick order a horseshoe
The resource pick order simply went in a circle. It was originally
implemented this way for ease, but the rules state it should go forward
for common and then backwards (starting with the last player) for rares.
The pick order now correctly follows the rules.

There was also a bug where if the first drawn omen was a war card the
player would not be offered the opportunity to attack. This wasn't really
a problem since an attack on turn 1 can't do anything anyway. I fixed it
so that it doesn't become a real problem if there are any rules changes.

I did a substantial amount of code refactoring while I was in the resource
picking functions. Now most of their logic is moves into common helper
functions at the top of the file. These helper functions modifiy the game
object, but do not persist changes. The resource picking functions are
responsible for persisting changes to the game state at the end. Since
changes are only persisted once and the whole game object is saved, the
changes are now atomic.