Style game page except modal actions
The rest of the game play page is not styles including structures, the
marketplace, shared tracks, weregeld, decks, etc. The cheat section was
moved far down the page. The only section remaining to style are the modal
actions like raiding or collecting bounty which are unstlyed and on a
random part of the page.